using CGFloat for height blows up core data on iPhone 6 and above since core data weill no longer force the coercion on those models.  Using float seems to work ok as UIView seems to be ok with it
coerced longs to int for storage in a few places since we never store bignumbers

+++ IcyBee/IcyBee/ChatMessage.h
@@ -16,7 +16,8 @@
 @property (nonatomic, retain) NSDate    *timeStamp;
 @property (nonatomic, retain) NSString  *sender;
 @property (nonatomic, retain) NSString  *text;
-@property (nonatomic)         CGFloat   height;
+//@property (nonatomic)         CGFloat   height;
+@property (nonatomic)         float     height;
 @property (nonatomic)         BOOL      url;
 @property (nonatomic)         BOOL      needsSize;
 @property (nonatomic)         int       groupIndex;
